[ashell] Minor fixes and cleanup
